requirements.txt
ultralytics>=8.3
gradio>=5.0
flask>=3.0
opencv-python>=4.8
Pillow>=10.0
